requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
coverage==5.3
CurrencyConverter==0.16.4
dal-admin-filters==1.0.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.3
django-autocomplete-light==3.8.1
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==2.4.0
django-js-asset==1.2.2
django-simple-autocomplete==1.11
django-test-plus==1.4.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
Faker==0.9.1
filelock==3.0.12
forex-python==1.5
idna==2.10
importlib-metadata==3.3.0
importlib-resources==3.3.0
iniconfig==1.1.1
instagram-basic-display==1.0.5
jdcal==1.4.1
jsonfield==2.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.8
pbr==5.5.1
# Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
razorpay==1.2.0
requests==2.25.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
stevedore==3.3.0
tablib==2.0.0
text-unidecode==1.2
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
